SiteUser 엔티티 추가, 다른 엔티티의 주키를 Long타입으로

+++ src/main/java/com/exam/sbb/answer/AnswerController.java
@@ -20,7 +20,7 @@ public class AnswerController {
     private final AnswerService answerService;
 
     @PostMapping("/create/{id}")
-    public String createAnswer(Model model, @PathVariable int id, @Valid AnswerForm answerForm, BindingResult bindingResult){
+    public String createAnswer(Model model, @PathVariable Long id, @Valid AnswerForm answerForm, BindingResult bindingResult){
         Question question = questionService.getQuestion(id);
 
         if(bindingResult.hasErrors()){
